+++ src/renderer/styles/arco-override.scss
@@ -6,6 +6,7 @@
 // 该 token 同时影响 Card/Tag/Menu/Checkbox 等使用 small 圆角的组件。
 body {
   --border-radius-small: 8px;
+  --border-radius-medium: 10px;
 
   // Arco 边框 token 桥接：Arco 组件在 body 上消费 --color-border-2（写到 :root 会被 body 遮蔽），
   // 故在此覆盖为项目语义 token。值引用 --border-light，自动随主题切换。
@@ -312,12 +313,6 @@ body[arco-theme='dark'] .arco-message {
     height: 2px !important;
     border-radius: 999px !important;
   }
-}
-
-.arco-select-view,
-.arco-input,
-.arco-input-inner-wrapper {
-  background-color: var(--fill-0) !important;
 }
 
 // ---------- Skill Hub input / select ----------
